+++ week3-recommend-menu/src/App.jsx
@@ -1,13 +1,13 @@
 import { ThemeProvider } from "styled-components";
 import GlobalStyle from "./styles/GlobalStyle";
 import theme from "./styles/theme";
-import MainHome from "./components/MainHome";
+import MainLayout from "./components/MainLayout";
 
 function App() {
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
-      <MainHome />
+      <MainLayout />
     </ThemeProvider>
   );
 }
